Extract HorizontalScroll into a sepparate component
